+++ src/workspaces/NewWorkspaceWizard/NewWorkspaceWizard.tsx
@@ -17,8 +17,8 @@ import Events, { extractCrossWorkspaceDetails, extractWorkspaceDetails } from 's
 import * as Nav from 'src/libs/nav';
 import { useCancellation, useOnMount, withDisplayName } from 'src/libs/react-utils';
 import * as Utils from 'src/libs/utils';
-import { CloneEgressWarning } from 'src/workspaces/NewWorkspaceModal/CloneEgressWarning';
-import { CreatingWorkspaceMessage } from 'src/workspaces/NewWorkspaceModal/CreatingWorkspaceMessage';
+import { CloneEgressWarning } from 'src/workspaces/NewWorkspaceWizard/CloneEgressWarning';
+import { CreatingWorkspaceMessage } from 'src/workspaces/NewWorkspaceWizard/CreatingWorkspaceMessage';
 import { isGoogleWorkspace, isProtectedWorkspace, WorkspaceInfo, WorkspaceWrapper } from 'src/workspaces/utils';
 import validate from 'validate.js';
