+++ src/workspaces/list/WorkspacesListModals.test.tsx
@@ -5,17 +5,17 @@ import { renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
 import { notifyNewWorkspaceClone } from 'src/workspaces/common/state/useCloningWorkspaceNotifications';
 import { WorkspaceUserActionsContext } from 'src/workspaces/list/WorkspaceUserActions';
-import { NewWorkspaceWizardProps } from 'src/workspaces/NewWorkspaceModal//NewWorkspaceWizard';
+import { NewWorkspaceWizardProps } from 'src/workspaces/NewWorkspaceWizard/NewWorkspaceWizard';
 import { WorkspaceInfo } from 'src/workspaces/utils';
 
 const mockModalFn = jest.fn();
 
-type NewWorkspaceWizardExports = typeof import('src/workspaces/NewWorkspaceModal/NewWorkspaceWizard') & {
+type NewWorkspaceWizardExports = typeof import('src/workspaces/NewWorkspaceWizard/NewWorkspaceWizard') & {
   __esModule: true;
 };
-jest.mock('src/workspaces/NewWorkspaceModal/NewWorkspaceWizard', (): NewWorkspaceWizardExports => {
+jest.mock('src/workspaces/NewWorkspaceWizard/NewWorkspaceWizard', (): NewWorkspaceWizardExports => {
   return {
-    ...jest.requireActual<NewWorkspaceWizardExports>('src/workspaces/NewWorkspaceModal/NewWorkspaceWizard'),
+    ...jest.requireActual<NewWorkspaceWizardExports>('src/workspaces/NewWorkspaceWizard/NewWorkspaceWizard'),
     default: mockModalFn,
     __esModule: true,
   };
